log most recent resize time in main.js
I think there is a bug in the project code for timing of resizing pizzas. Line 466 says:
console.log("Time to resize pizzas: " + timeToResize[0].duration + "ms");

which works fine for the first resize event, but additional resizes go to the end of the timeToResize array, not the beginning. That is why all resizes give the same duration, which would be rather shocking if it were true. The code should probably be something like:

console.log("Time to resize pizzas: " + timeToResize[timeToResize.length-1].duration + "ms");